requirements.txt
Flask
gunicorn
flask_login
supabase
flask_sqlalchemy
Pillow
PyMuPDF